Added .Length() function to each class. Added const Pi
the .Length() function can be used directly on a Vector type, instead of calling Length(Vector type), it's much easier, trust me. Added the const pi to the length of 11 decimals, should be accurate enough. It's called through ZCPP.Pi, but if you have the namespace in the ZCPP. part wont be necessary.